Move ManagedStackError class to the module 'managed_cloudformation_stack'
